Alt prevents or releases Shift
Makes "Alt + Shift + " key combinations possible.

There is a catch though. If you press Shift before Alt, it's going to release the Shift which is interpreted by calc as some type of "One time shift lock". So:
Shift + Alt + 9 = Shift + (
Alt + Shift + 9 = (

I don't know if nspire-io exhibits the same behavior, but it is something to keep in mind when testing.