[x86_64 JIT] Support jumps/calls beyond a 32bit relative offset
Currently the JIT emits relative jumps/calls when calling into
asmcode_x86_64.S, but that only works if the JIT code area is close enough
in memory. So far this was managed by not allowing the final executable to
be position-independent and forcing the JIT code area to be placed at
0x30000000.

Implement support for 64bit absolute calls by emitting a table of addresses
which can be used by rip-relative indirect jump/call instructions.

This means PIC/PIE code is supported now, so drop the -no-pie enforcement
and let mmap pick an appropriate location.