Bump. (#2)
* Use only -no-pie/-nopie instead of combining it with -fno-pie

- Resulted in "-fPIC -fno-pie" beforehand, which results in broken code
  causing "signal not found" when built with GCC

* Only the x86 and x86_64 JITs do not support PIE

* Move speedup hack from addr_cache_flush() to coproc.cpp

- Otherwise the cache flush is also skipped if firebird needs it

* Remove unused debugger_input global

* Don't use LTO with clang on Linux

- Disabling LTO for just the UI does not seem to work with qmake, so just
  disable it completely with clang.

* Remove some code from firebird.pro which is not necessary anymore

- QMAKE_CXXFLAGS has sane default values now after the fix for android clang
- LTO for iOS appears to work meanwhile

* [AArch64 JIT] Fix typo

* Fix Primecell ID of serial_cx

- The datasheet has two different values, the one on the summary page is wrong

* Fix usblink_queue_send_os definition

- Did not match declaration

* Bump Qt min version to 5.9

* Remove unused exec_hack()

RF_EXEC_HACK got dropped a while ago as well

* Don't use deprecated QString::vsprintf

* Refactor use of setjmp/longjmp

- Use __builtins only when necessary
- Make restart_after_exception static
- Drop NO_SETJMP define

* Make gzip_filesize static

* Fully optimize final binary on Linux with g++

- Qt passes -Wl,01 to g++ when linking, which is not good for LTO

* Fix android build with Qt 5.14.1 and add arm64-v8a support

- Changes in firebird.pro needed for dealing with multiple ANDROID_ABIS
  as supported by Qt 5.14+
- Only the clang assembler for iOS seems to need the @PAGE/@PAGEOFF syntax
- Update the build.gradle file to work with gradle > 2.2

* No x86 or x86_64 JIT on android

* Disable .text relocations on armeabi-v7a android

* Support native Android file picker used by Qt >= 5.13

- Qt 5.13 uses an Intent for file picking now, returning a content:// URL
- Add some hacks to be able to display and handle those URLs
- Add a hack to use the open dialog for save dialogs as well

* macOS: don't use MAP_32BIT for os_alloc_executable's mmap.

This made the JIT not work anymore, at least on macOS 10.15.
Would be nice to test this on older OSes now...

* iOS: use ptrace PT_DENY_ATTACH in translate_deinit. Fixes #124

* iOS: W^X mmap for JIT. mprotect will take care of it later.

* Fix some warnings:

- Avoid implicit signedness change in addr_cache_invalidate
- FASTCALL needs a full prototype, use (void) explicitly
- Remove unused ac_commit_map and boot2_base

* Close sockets on MinGW properly, use -1 consistently

Using close for sockets on MinGW is not possible, it generated an error
message and the socket was left open.

* [GUI] Fix monospace font not always being monospace.

* iOS: JIT: make sure to call ptrace(PT_DENY_ATTACH) at exit.

* Bump to version 1.5

* Bump to 1.5-dev

* Drop archlinux/PKGBUILD

It doesn't really make sense to maintain that here, it's on OBS and in the AUR
already.

* Better management of build-time configuration options

- Allow setting TRANSLATION_ENABLED and SUPPORT_LINUX to "auto"
- Print the value of relevant variables during qmake

* Fix disabling x86/x86_64 JIT on android

- One ')' too many
- Condition was always false
- How did qmake not complain and how did it change the evaluation?

* Make the fake Firebird.Emu.Emu more complete

* Redo MobileUIFront with Qt Quick Layouts

It's now just a single GridLayout with minimal changes required for
(left handed) portrait mode.

Also drop the bar with actions, the drawer made it redundant.

* [MobileUI] Size the drawer based on its contents

Now items are no longer cut off and it's not bigger than necessary.

* Add Polish translation

Signed-off-by: Atrate <Atrate@protonmail.com>

* Capitalize Firebird where appropriate

Signed-off-by: Atrate <Atrate@protonmail.com>

* Add generated Polish translation file

Signed-off-by: Atrate <Atrate@protonmail.com>

* Add pl_PL.qm to resources.qrc

Signed-off-by: Atrate <Atrate@protonmail.com>

* Add lrelease embed_translation to CONFIG

Signed-off-by: Atrate <Atrate@protonmail.com>

* Revert "Add lrelease embed_translation to CONFIG"

This reverts commit ad628185ac1537b996ecce415fb0b775cf45e496.

* [AArch64 JIT] Support negative offset in literalpool_fill

Not used currently, but avoid blowing up later.

* [ARM JIT] Fix comment

* [x86_64 JIT] Support jumps/calls beyond a 32bit relative offset

Currently the JIT emits relative jumps/calls when calling into
asmcode_x86_64.S, but that only works if the JIT code area is close enough
in memory. So far this was managed by not allowing the final executable to
be position-independent and forcing the JIT code area to be placed at
0x30000000.

Implement support for 64bit absolute calls by emitting a table of addresses
which can be used by rip-relative indirect jump/call instructions.

This means PIC/PIE code is supported now, so drop the -no-pie enforcement
and let mmap pick an appropriate location.

* Remove unused variable in os-linux.c, use correct condition for 31bit alloc

mem_and_flags only needs to have bit 31 clear if the x86 format is used,
which is decided in mmu.h. Use the proper condition.

* Fix -Waddress-of-packed-member warnings

`warning: taking address of packed member of 'arm_state' may result in an unaligned pointer value`

The packed attribute also sets the alignment of the struct to 1, but that's
not actually wanted at all. Avoid that effect by explicitly specifying the
alignment later.

* Drop mention of linux-g++-32 in firebird.pro

It's used for cross-compiling with qmake, but what this section does is already
handled by qmake.

* Switch to bionic and enable arm64 and ppc64le builds in .travis.yml

* Speed up GPSP significantly

It spent most of the CPU time in the LCD polling loop.
Also copy the same speedup from timer_read to timer_cx_read.

* Implement fastboot RAM properly

90030000 appears to be just 4KiB of RAM.
Also rename the meanwhile known SPI controller on CX at 90040000.

With this, the installer 4.5.3.14 for CX succeeds.

* fix strncasecmp include.

* emscripten: fix build.

- Doesn't work with closure anymore (probably because WASM is not supported)
- Some functions have to be exported manually
- TOTAL_MEMORY has to be a multiple of 64KB

* emscripten: clean up HTML.

A whole lot of things aren't needed anymore.

Co-authored-by: Fabian Vogt <fabian@ritter-vogt.de>
Co-authored-by: Adrien Bertrand <bertrand.adrien@gmail.com>
Co-authored-by: Atrate <Atrate@protonmail.com>