Don't ignore other faults in mmu_check_priv
Previously, it only aborted in the case of a section permission fault (0xD),
but it also has to check for page permission faults (0xF).
However, there's no reason to ignore any kind of fault here, as there's an
actual access to the address afterwards, which would abort anyway. So just
handle all faults here and rename the function accordingly.