New flash creation dialog in QML
A rewritten and slightly redesigned flash creation dialog in QML.
This doesn't depend on QWidgets and can thus be used on Android/iOS as well.

Some functions to read metadata from flash components were added to the core,
they work in a less hacky way than the ones in flashdialog.cpp previously.

Fixes #135