Note additional conflicts in keymappings
Noticed some mappings that, at least on Windows boxen, are in conflict with the keyboard shortcuts for the 'Emulation' and 'Flash' menubar items.

For reference to any reading this that are unfamiliar, if a letter in a menubar item is underlined, then pressing the [Alt] key while also pressing the key of the underlined letter will cause that particular menuitem to expand, highlighting the first item in its submenu. Firebird has the following shortcuts of this type configured: [E]->'Emulation', [T]->'Tools', [N]->'Snapshot', [F]->'Flash', and [U]->'About'. The 'Docks' menubar item does not have a shortcut assigned to it.

Fortunately, both of the conflicts have alternative mappings as well, so removing these lines is of minimal negative impact.

Unfortunately, this person has no suggestions for a better mapping, so changing these lines rather than removing them may be difficult.