Handle all TDs on USB transmit (calc -> host)
Previously only the first TD was looked at when an EP got primed. However,
if that TD didn't emit an interrupt, the transfer got stuck as nothing
advanced the queue. This happens in the case of size % maxpacket == 0, when
a zero length packet is sent after the data transfer. The zero length
packet is sent with the next TD in the queue and has the interrupt on
complete set. Loop through all TDs in the queue until the terminate bit
is set, so that all TDs are handled.