[CX II] Implement sleep mode
With this, it's possible to turn the calculator back on after it went into
sleep mode (idle timeout or explicitly turned off).

What happens when pressing On is that the calc starts from bootrom again, but
that finds fastboot data in the 90030000 area left by the OS and uses that to
validate some state and jumps back to the OS code in SDRAM.

Add a TODO entry to improve fastboot data handling. Currently the data is
preserved across restarts, which means that some possibly incompatible bootrom
tries to interpret it. So far the validation detects that SDRAM was cleared and
aborts.